feat(common, broker): #WB2-2701, dynamic routes using constructor params in impl
