--- broker-parent/broker-client/ts/src/nats/directory-group-member-add.client.ts
@@ -1,39 +1,0 @@
-import { Inject, Injectable } from '@nestjs/common';
-import { ClientProxy } from '@nestjs/microservices';
-import { firstValueFrom } from 'rxjs';
-import { AddGroupMemberRequestDTO, AddGroupMemberResponseDTO } from './directory-group-member-add.types';
-
-@Injectable()
-export class DirectoryGroupMemberAddClient {
-
-  constructor(
-    @Inject('NATS_CLIENT') private readonly natsClient: ClientProxy
-  ) {
-    console.log('Creating events service')
-  }
-
-  
-    
-  async request(event: AddGroupMemberRequestDTO): Promise<AddGroupMemberResponseDTO> {
-    const eventAddress = this.getSubject('directory.group.member.add');
-    const reply = await firstValueFrom(this.natsClient.send(eventAddress, event));
-    return reply as AddGroupMemberResponseDTO;
-  }
-        
-
-
-  private getSubject(subjectPattern: string, ...addressComplements: string[]): string {
-    let subject: string;
-    if(addressComplements) {
-      let parts = subjectPattern.split(/[*>]/);
-      for(let i = 0; i < addressComplements.length; i++) {
-        parts.splice(2 * i + 1, 0, addressComplements[i]);
-      }
-      subject = parts.join('');
-    } else {
-      subject = subjectPattern;
-    }
-    return subject;
-  }
-
-}
