--- broker-parent/broker-client/ts/src/nats/directory-group-member-add.types.ts
@@ -1,11 +1,0 @@
-export interface AddGroupMemberRequestDTO {
-  groupId?: string;
-  groupExternalId?: string;
-  userId?: string;
-}
-
-
-export interface AddGroupMemberResponseDTO {
-  added?: boolean;
-}
-
