--- broker-parent/broker-client/ts/src/nats/ent-test-listen-reply.types.ts
@@ -1,12 +1,0 @@
-export interface ListenAndAnswerDTO {
-  userId?: string;
-  jobId?: string;
-}
-
-
-export interface DummyResponseDTO {
-  userId?: string;
-  jobId?: string;
-  success?: boolean;
-}
-
