--- broker-parent/broker-client/ts/src/nats/ent-test-listen.types.ts
@@ -1,7 +1,0 @@
-export interface ListenOnlyDTO {
-  userId?: string;
-  timestamp?: number;
-}
-
-
-
