--- broker-parent/broker-client/ts/src/nats/i18n-translations-fetch.types.ts
@@ -1,22 +1,0 @@
-export interface FetchTranslationsRequestDTO {
-  headers?: {
-    [k: string]: {
-      [k: string]: unknown;
-    };
-  };
-  langAndDomain?: LangAndDomain;
-}
-export interface LangAndDomain {
-  lang?: string;
-  domain?: string;
-}
-
-
-export interface FetchTranslationsResponseDTO {
-  translations?: {
-    [k: string]: {
-      [k: string]: unknown;
-    };
-  };
-}
-
